requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
redis
pytest
